add mode argument to CONNECT command rename usb.py to usb_transport.py to avoid naming conflicts with the pyusb library
