+++ pyxcp/__init__.py
@@ -31,7 +31,6 @@ __copyright__ = """
   with this program; if not, write to the Free Software Foundation, Inc.,
   51 Franklin Street, Fifth Floor, Boston, MA 02110-1301 USA.
 """
-
 from .version import __version__
 
 VERSION = sys.version_info
